gpiops:MISRA-C:Literal value requires a U suffix.
This patch fixes below MISRA-C violation in gpiops
Literal value requires a U suffix.
MISRA-C:2012/AMD1 R.7.2,R.10.3,R.10.4


Acked-for-series: Srinivas Goud <srinivas.goud@xilinx.com>

+++ XilinxProcessorIPLib/drivers/gpiops/src/xgpiops.c
@@ -139,7 +139,7 @@ s32 XGpioPs_CfgInitialize(XGpioPs *InstancePtr, const XGpioPs_Config *ConfigPtr,
 	 * By default, interrupts are not masked in GPIO. Disable
 	 * interrupts for all pins in all the 4 banks.
 	 */
-	for (i=0;i<InstancePtr->MaxBanks;i++) {
+	for (i=(u8)0U;i<InstancePtr->MaxBanks;i++) {
 		XGpioPs_WriteReg(InstancePtr->GpioConfig.BaseAddr,
 					  ((u32)(i) * XGPIOPS_REG_MASK_OFFSET) +
 					  XGPIOPS_INTDIS_OFFSET, 0xFFFFFFFFU);
